Executable container group reference name string
Each executable container need to have a grouping reference name, as for all different stages it can be "stages", or for multiple different rounds it can be "rounds". Default value for group reference name is assigned as EC type name + "s", going for plural. But any values can be specified in it.